Fix file-sort UI test across midnight
